pyproject.toml
[project]
name = "pro_1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
]
